requirements.txt
# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
SQLAlchemy==2.0.23
sqlite3

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Image Processing
Pillow==10.1.0

# Environment
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1

# DateTime
dateutils==0.6.12

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Development
Werkzeug==3.0.1
click==8.1.7